+++ core/src/main/java/com/admanager/core/BannerLoader.java
@@ -108,10 +108,10 @@ public abstract class BannerLoader<L extends BannerLoader> {
     }
 
     protected void clicked() {
-        clicked(null);
+        clicked(null, null);
     }
 
-    protected void clicked(final String subname) {
+    protected void clicked(final String subname, final Double revenue) {
         Log.v(TAG, getAdapterName() + ": clicked");
         if (clickListener == null) {
             return;
@@ -123,7 +123,7 @@ public abstract class BannerLoader<L extends BannerLoader> {
                 if (subname != null) {
                     name = name + "_" + subname;
                 }
-                clickListener.clicked(name);
+                clickListener.clicked(name, revenue);
             }
         });
     }
@@ -280,6 +280,6 @@ public abstract class BannerLoader<L extends BannerLoader> {
     }
 
     public interface ClickListener {
-        void clicked(String adapterName);
+        void clicked(String adapterName, Double revenue);
     }
 }